+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/NotificationDialogWithNoTitle.test.jsx
@@ -1,0 +1,35 @@
+import React, { useState } from 'react';
+import Button from 'terra-button';
+import NotificationDialog from '../../../NotificationDialog';
+
+const NotificationDialogNoTitle = () => {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const handleOpenModal = () => {
+    setIsOpen(true);
+  };
+
+  const handleCloseModal = () => {
+    setIsOpen(false);
+  };
+
+  return (
+    <>
+      {isOpen && (
+        <NotificationDialog
+          variant="error"
+          startMessage="The Main Instruction is text used to provide more detail or define terminology. Don’t repeat the title verbatim."
+          acceptAction={{
+            text: 'Close',
+            onClick: handleCloseModal,
+          }}
+          buttonOrder="acceptFirst"
+          emphasizedAction="accept"
+        />
+      )}
+      <Button id="trigger-notification-dialog" text="Trigger NotificationDialog" onClick={handleOpenModal} />
+    </>
+  );
+};
+
+export default NotificationDialogNoTitle;