+++ packages/terra-dialog-modal/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Bump terra-notification-dialog version from ^3 to ^4
+
 ## 3.45.0 - (August 18, 2020)
 
 * Changed
@@ -420,5 +423,4 @@
 
 ## 1.0.0 - (June 29, 2018)
 
-* Added
 * Initial stable release